fix unsigned long long
git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@428 a60ba226-48e3-df11-83e3-000476a39db3

+++ include/optimizer/Intervals.h
@@ -5,7 +5,7 @@
 
 class Interval {
 public:
-    typedef long Integer;
+    typedef unsigned long long Integer;
 
     // construct ∅
     Interval();
